+++ es6/es6.spec.js
@@ -43,3 +43,27 @@ test('security/detect-child-process', async () => {
     1 problem"
   `);
 });
+
+test('no console', async () => {
+  expect(await getErrors('no-console.js')).toMatchInlineSnapshot(`
+    "__fixtures__/no-console.js: line 1, col 1, Error - Unexpected console statement. (no-console)
+
+    1 problem"
+  `);
+});
+
+test('no alert', async () => {
+  expect(await getErrors('no-alert.js')).toMatchInlineSnapshot(`
+    "__fixtures__/no-alert.js: line 1, col 1, Error - Unexpected alert. (no-alert)
+
+    1 problem"
+  `);
+});
+
+test('no debugger', async () => {
+  expect(await getErrors('no-debugger.js')).toMatchInlineSnapshot(`
+    "__fixtures__/no-debugger.js: line 1, col 1, Error - Unexpected 'debugger' statement. (no-debugger)
+
+    1 problem"
+  `);
+});