+++ es6/rules.js
@@ -6,6 +6,11 @@ module.exports = {
     // Allow unused args with underscore
     'no-unused-vars': ['error', { argsIgnorePattern: '^_' }],
 
+    // no console
+    'no-console': 'error',
+    'no-alert': 'error',
+    'no-debugger': 'error',
+
     // conflicts with prettier
     'array-bracket-spacing': 'off',
     'max-len': 'off',