+++ CHANGELOG.md
@@ -2,6 +2,18 @@ CHANGELOG
 =========
 
 ## HEAD (Unreleased)
-_(none)_
+
+* Major: added `'no-console': 'error'`
+* Major: added `'no-alert': 'error'`
+* Major: added `'no-debugger': 'error'`
+* Patch: added tests
 
 ---
+
+## 2.0.0 (2021-01-13)
+* Major: Enable `@typescript-eslint/naming-convention`
+* Minor: Allow unused args prefixed with underscore
+* Minor: Fix `@typescript-eslint/no-shadow`
+* Minor: Fix `@typescript-eslint/no-non-null-assertion`
+* Minor: Disable `security/detect-non-literal-fs-filename`
+